# docs/swagger.yaml
definitions:
  auth.Authuser:
    properties:
      address:
        type: string
      nonce:
        type: string
    type: object
  auth.SigninPayload:
    properties:
      address:
        type: string
      nonce:
        type: string
      sig:
        type: string
    type: object
...
host: restwalletchat-app-sey3k.ondigitalocean.app
info:
  contact:
    email: walletchatextension@gmail.com
    url: https://walletchat.fun
  description: |
    Wecome to the WalletChat API Documentation

    Please make note that some JSON data structures are shared for both input/output.
    Required input parameters will have a red * next to them in the data type outline at
    the bottom of the page, along with a comment.  This means when executing API functionality
    from this API page, some fields may need to be removed from the JSON struct before submitting.
    Please email the developers with any issues.
    Some JSON data structures are output only, and will be marked as such as well.

    v0 of the API does not include encryption or authentication.  Please as you are given access
    to this page, do not abuse this system and impersonate others, or submit offensive material.
    developers monitor this data daily.

    v1 will include basic JWT Authentication, however some additional work is in progress to make this fully secure.
    except for AUTH functions, all endpoints must prefix /v1 and include "Bearer: <JWT>" in all reqests"

    v2 will include encyrption for DMs, private keys will be stored locally on client PCs
    with no way for us to recover any data which is encrypted.
  title: WalletChat API
  version: "1.0"
paths:
...
  /register:
    post:
      consumes:
      - application/json
      description: |-
        This is a one-time operation, maybe could be combined into the nonce-generating call.  Basically places a wallet address
        into the database for further use.  Only the "address" field is needed for input here in the AuthUser struct.
      parameters:
      - description: address input in json format
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/auth.Authuser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Register Wallet Address for the first time, one-time operation
      tags:
      - Auth
  /signin:
    post:
      consumes:
      - application/json
      description: |-
        Every call the to API after this signin should present the JWT Bearer token for authenticated access.
        Upon request we can change the timeout to greater than 24 hours, or setup an addtional dedicated API for
        an agreed upon development and maintenance cost
      parameters:
      - description: json input containing signed nonce
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/auth.SigninPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Sign In with signed nonce value, currently JWT token returned should
        be valid for 24 hours
      tags:
      - Auth
...
  /users/{address}/nonce:
    get:
      consumes:
      - application/json
      description: |-
        As part of the login process, we need a user to sign a nonce genrated from the API, to prove the user in fact
        the owner of the wallet they are siging in from.  JWT currently set to 24 hour validity (could change this upon request)
      parameters:
      - description: wallet address to get nonce to sign
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      summary: If the current wallet doesn't have a valid local JWT, need to request
        a new nonce to sign
      tags:
      - Auth
swagger: "2.0"